data-backup/requirements.txt
async-timeout==4.0.2
boto3==1.27.0
botocore==1.30.0
dnspython==2.3.0
jmespath==1.0.1
pymongo==4.4.0
python-dateutil==2.8.2
redis==4.6.0
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16